refactor: move a function to pacquet-fs
