+++ crates/store-dir/Cargo.toml
@@ -11,6 +11,8 @@ license.workspace    = true
 repository.workspace = true
 
 [dependencies]
+pacquet-fs = { workspace = true }
+
 derive_more = { workspace = true }
 miette      = { workspace = true }
 serde       = { workspace = true }
